requirements.txt
requests==2.31.0
dnspython==2.4.2
colorama==0.4.6
Pillow==9.4.0
